chapter7-9/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-paypal==0.4.1
idna==2.6
kombu==4.1.0
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
